@@ -1,20 +1,10 @@
-#include <cassert>
-#include <string>
-#include <cstring>
-#include <cstdio>
-#include <cassert>
+#include <iostream>
 #include <string>
 #include <vector>
-#include <cstring>
-#include <iostream>
-#include <cstring>
+#include <algorithm>
+#include <cstdlib>
 
-#include "./llam_cpp/include/gpt4all-backend/llmodel.h"
-
-std::string answer = "";
-LLModel::PromptContext prompt_context;
-LLModel* model = nullptr;
-bool _stopFlag = false;
+#include <gpt4all-backend/llmodel.h>
 
 // --------- Struct for Parameters --------------
 struct llm_params {
@@ -35,26 +25,35 @@ struct llm_params {
     int number_of_gpu_layers = 80;
 };
 
-// --------- Argument Parser -------------------
+// --------- Global Variables -------------------
+std::string answer = "";
+// LLModel::PromptContext prompt_context;
+LLModel* model = nullptr;
+std::atomic<bool> _stopFlag = false;
+
+// --------- Print Help Message -----------------
 void llm_print_usage(const llm_params& p, const char* progname) {
     std::cout <<
-        std::string("Usage: %1 [options]").arg(progname) << "\n\n"
-                    "Options:\n"
-                    "  --stream [true|false]              Enable/disable streaming output (default: true)\n"
-                    "  --prompt-template <string>         Prompt template with %1 as input slot\n"
-                    "  --system-prompt <string>           System prompt text\n"
-                    "  --temperature <float>              Temperature (default: 0.7)\n"
-                    "  --top-k <int>                      Top-K sampling (default: 40)\n"
-                    "  --top-p <float>                    Top-P sampling (default: 0.4)\n"
-                    "  --min-p <float>                    Minimum probability (default: 0.0)\n"
-                    "  --repeat-penalty <float>           Repeat penalty (default: 1.18)\n"
-                    "  --prompt-batch-size <int>          Prompt batch size (default: 128)\n"
-                    "  --max-tokens <int>                 Max tokens (default: 4096)\n"
-                    "  --repeat-penalty-tokens <int>      Number of tokens for repetition penalty (default: 64)\n"
-                    "  --context-length <int>             Context length (default: 2048)\n"
-                    "  --gpu-layers <int>                 Number of GPU layers (default: 80)\n";
+        "Usage: " << progname << " [options]\n\n"
+                                 "Options:\n"
+                                 "  --model <path>                     Path to model file (default: models/model.gguf)\n"
+                                 "  --backend <cpu|cuda|auto>          Backend to use (default: cpu)\n"
+                                 "  --stream [true|false]              Enable streaming output (default: true)\n"
+                                 "  --prompt-template <string>         Prompt template with %1 as input slot\n"
+                                 "  --system-prompt <string>           System prompt text\n"
+                                 "  --temperature <float>              Temperature (default: 0.7)\n"
+                                 "  --top-k <int>                      Top-K sampling (default: 40)\n"
+                                 "  --top-p <float>                    Top-P sampling (default: 0.4)\n"
+                                 "  --min-p <float>                    Minimum probability (default: 0.0)\n"
+                                 "  --repeat-penalty <float>           Repeat penalty (default: 1.18)\n"
+                                 "  --prompt-batch-size <int>          Prompt batch size (default: 128)\n"
+                                 "  --max-tokens <int>                 Max tokens (default: 4096)\n"
+                                 "  --repeat-penalty-tokens <int>      Tokens for repetition penalty (default: 64)\n"
+                                 "  --context-length <int>             Context length (default: 2048)\n"
+                                 "  --gpu-layers <int>                 Number of GPU layers (default: 80)\n";
 }
 
+// --------- Argument Parser -------------------
 bool llm_parse_args(int argc, char* argv[], llm_params& params) {
     for (int i = 1; i < argc; ++i) {
         std::string arg = argv[i];
@@ -95,37 +94,78 @@ bool llm_parse_args(int argc, char* argv[], llm_params& params) {
         } else if (arg == "--gpu-layers") {
             params.number_of_gpu_layers = std::stoi(argv[++i]);
         } else {
-            std::cerr << "Unknown argument: " << arg << std::endl;
+            std::cerr << "Unknown argument: " << arg << "\n";
             llm_print_usage(params, argv[0]);
-            exit(0);
+            return false;
         }
     }
     return true;
 }
 
 // --------- Response Handler ------------------
-bool handleResponse(int32_t token, const std::string &response) {
-    const char* responsechars = response.c_str();
-    if (!(responsechars == nullptr || responsechars[0] == '\0')) {
-        std::cout << responsechars << std::flush;
-        answer += responsechars;
+bool handleResponse(int32_t token, std::string_view response) {
+    if (!response.empty()) {
+        std::cout << response << std::flush;
+        answer += std::string(response); // Convert string_view to string to append to answer
     }
     return !_stopFlag.load();
 }
 
+// --------- Generate Full Prompt and Process Request ----------------
+void processPrompt(const std::string& input, const llm_params& params) {
+    // Create full prompt by combining system prompt and user input
+    std::string full_prompt = params.system_prompt;
+    std::string user_prompt = params.prompt_template;
 
-int main(int argc, char *argv[]) {
+    // Replace "%1" in the template with the user's input
+    size_t pos = user_prompt.find("%1");
+    if (pos != std::string::npos) {
+        user_prompt.replace(pos, 2, input);
+    }
+
+    full_prompt += user_prompt;  // Combine system and user prompt
+
+    // Callback for the prompt (this is a basic implementation)
+    auto prompt_callback = [](std::span<const LLModel::Token> batch, bool cached) -> bool {
+        // Optionally handle the batch here
+        return true;
+    };
+
+    // Callback for receiving the response from the model
+    auto response_callback = [](int32_t token_id, std::string_view response) -> bool {
+        return handleResponse(token_id, response);
+    };
+
+    // Create and set prompt context
+    LLModel::PromptContext ctx;
+    ctx.n_predict = params.max_tokens;  // Adjust as needed
+    ctx.top_k = params.top_k;
+    ctx.top_p = params.top_p;
+    ctx.min_p = params.min_p;
+    ctx.temp = params.temperature;
+    ctx.n_batch = params.prompt_batch_size;
+    ctx.repeat_penalty = params.repeat_penalty;
+    ctx.repeat_last_n = params.repeat_penalty_tokens;
+    ctx.contextErase = 0.5f;  // Customize if needed
+
+    // Call the prompt method on the model
+    model->prompt(full_prompt, prompt_callback, response_callback, ctx);
+
+    std::cout << "\n__DONE__\n" << std::flush;
+}
+
+// --------- Main Function ---------------------
+int main(int argc, char* argv[]) {
     llm_params params;
 
     if (!llm_parse_args(argc, argv, params))
         return 1;
 
-    prompt_context.n_ctx = params.context_length;
-    prompt_context.n_predict = params.max_tokens;
+    // prompt_context.n_predict = params.max_tokens;
 
-    model = LLModel::Implementation::construct(params.model, params.backend, prompt_context.n_ctx);
+    model = LLModel::Implementation::construct(params.model, params.backend, params.context_length);
     if (!model) {
-        qWarning() << "Failed to construct model.";
+        std::cerr << "Failed to construct model.\n";
         return 1;
     }
 
@@ -145,42 +185,34 @@ int main(int argc, char *argv[]) {
     }
 #endif
 
-    if (!model->loadModel(params.model, prompt_context.n_ctx, params.number_of_gpu_layers)) {
-        std::cout  << "Model loading unsuccessful!\n";
+    if (!model->loadModel(params.model, params.context_length, params.number_of_gpu_layers)) {
+        std::cerr << "Model loading unsuccessful!\n";
         return 1;
     } else {
-        std::cout  << "Model loaded successfully!\n";
+        std::cout << "Model loaded successfully!\n";
     }
 
-    bool chatInProcess = true;
+    while (true) {
+        std::string command;
+        std::getline(std::cin, command);
 
-    while(chatInProcess){
-        std::string line;
-        std::getline(std::cin, line);
-
-        std::string input;
-
-        if (line == "__STOP__") {
+        if (command == "__EXIT__") break;
+        if (command == "__STOP__") {
             _stopFlag.store(true);
             continue;
-        }else if (line == "__EXIT__") {
-            break;
-        }else if (line == "__PROMPT__") {
-            std::getline(std::cin, input);
         }
 
-        _stopFlag.store(false);
-        answer.clear();
+        if (command == "__PROMPT__") {
+            std::string input;
+            std::getline(std::cin, input);
 
-        auto prompt_callback = [](int32_t token_id) { return true;};
-        auto response_callback = std::bind(&handleResponse, this, std::placeholders::_1, std::placeholders::_2);
-        auto recalculate_callback = [](bool is_recalculating) {return is_recalculating;};
+            _stopFlag.store(false);
+            answer.clear();
 
-        model->prompt( input , params.prompt_template, prompt_callback, response_callback, recalculate_callback,prompt_context , false, nullptr);
-
-        std::cout << "\n__DONE__\n" << std::flush;
-
+            // Process and send the prompt to the model
+            processPrompt(input, params);
+        }
     }
 
-    return a.exec();
+    return 0;
 }
